created class TimeSeriesGenerator, moved statistics.py from models module to utils module, adjusted other files to accomodate changes
