+++ u8timeseries/models/theta.py
@@ -4,7 +4,7 @@ from ..custom_logging import raise_log, time_log, get_logger
 from .autoregressive_model import AutoRegressiveModel
 import numpy as np
 import math
-from .statistics import check_seasonality, extract_trend_and_seasonality, remove_seasonality
+from ..utils.statistics import check_seasonality, extract_trend_and_seasonality, remove_seasonality
 
 logger = get_logger(__name__)
