+++ u8timeseries/tests/test_timeseries_generation.py
@@ -3,14 +3,7 @@ import pandas as pd
 import numpy as np
 
 from ..timeseries import TimeSeries
-from ..utils.timeseries_generation import (
-    constant_timeseries, 
-    linear_timeseries, 
-    sine_timeseries,
-    gaussian_timeseries, 
-    random_walk_timeseries, 
-    us_holiday_timeseries
-)
+from ..utils.timeseries_generation import TimeSeriesGenerator as tg
 
 class TimeSeriesGenerationTestCase(unittest.TestCase):
 
@@ -21,7 +14,7 @@ class TimeSeriesGenerationTestCase(unittest.TestCase):
         value = 5
 
         # testing for constant value
-        constant_ts = constant_timeseries(value=value, length=length)
+        constant_ts = tg.constant_timeseries(value=value, length=length)
         value_set = set(constant_ts._series.values)
         self.assertTrue(len(value_set) == 1)
 
@@ -33,7 +26,7 @@ class TimeSeriesGenerationTestCase(unittest.TestCase):
         end_value = 12
 
         # testing for start value, end value and delta between two adjacent entries
-        linear_ts = linear_timeseries(start_value=start_value, end_value=end_value, length=length)
+        linear_ts = tg.linear_timeseries(start_value=start_value, end_value=end_value, length=length)
         self.assertEqual(linear_ts.values()[0], start_value)
         self.assertEqual(linear_ts.values()[-1], end_value)
         self.assertAlmostEqual(linear_ts.values()[-1] - linear_ts.values()[-2], (end_value - start_value) / (length - 1))
@@ -46,7 +39,7 @@ class TimeSeriesGenerationTestCase(unittest.TestCase):
         value_y_offset = -3
 
         # testing for correct value range
-        sine_ts = sine_timeseries(length=length, value_amplitude=value_amplitude, value_y_offset=value_y_offset)
+        sine_ts = tg.sine_timeseries(length=length, value_amplitude=value_amplitude, value_y_offset=value_y_offset)
         self.assertTrue((sine_ts <= value_y_offset + value_amplitude).all())
         self.assertTrue((sine_ts >= value_y_offset - value_amplitude).all())
 
@@ -56,7 +49,7 @@ class TimeSeriesGenerationTestCase(unittest.TestCase):
         length = 100
 
         # testing for correct length
-        gaussian_ts = gaussian_timeseries(length=length)
+        gaussian_ts = tg.gaussian_timeseries(length=length)
         self.assertEqual(len(gaussian_ts), length)
 
     def test_random_walk_timeseries(self):
@@ -65,7 +58,7 @@ class TimeSeriesGenerationTestCase(unittest.TestCase):
         length = 100
 
         # testing for correct length
-        random_walk_ts = random_walk_timeseries(length=length)
+        random_walk_ts = tg.random_walk_timeseries(length=length)
         self.assertEqual(len(random_walk_ts), length)
 
     def test_us_holiday_timeseries(self):
@@ -75,6 +68,6 @@ class TimeSeriesGenerationTestCase(unittest.TestCase):
         start_ts = pd.Timestamp('20201201')
 
         # testing for christmas and non-holiday
-        us_holiday_ts = us_holiday_timeseries(length=length, start_ts=start_ts)
+        us_holiday_ts = tg.us_holiday_timeseries(length=length, start_ts=start_ts)
         self.assertEqual(us_holiday_ts._series.at[pd.Timestamp('20201225')], 1)
         self.assertEqual(us_holiday_ts._series.at[pd.Timestamp('20201210')], 0)
